Fixed a mem leak. Restructured queueing. Simplified config.
